chore: Fix billing backend tests
