+++ ee/backend/app/src/main/kotlin/io/tolgee/ee/service/eeSubscription/usageReporting/ScheduledReportingManager.kt
@@ -1,17 +1,13 @@
 package io.tolgee.ee.service.eeSubscription.usageReporting
 
+import io.tolgee.component.SchedulingManager
 import io.tolgee.ee.EeProperties
 import io.tolgee.util.Logging
 import io.tolgee.util.logger
-import jakarta.annotation.PreDestroy
 import org.springframework.boot.context.event.ApplicationReadyEvent
 import org.springframework.context.event.EventListener
-import org.springframework.scheduling.TaskScheduler
 import org.springframework.stereotype.Component
 import java.time.Duration
-import java.util.*
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.ScheduledFuture
 
 /**
  * Manages scheduled reporting tasks using manual scheduling instead of @Scheduled annotation.
@@ -22,53 +18,27 @@ import java.util.concurrent.ScheduledFuture
  */
 @Component
 class ScheduledReportingManager(
-  private val taskScheduler: TaskScheduler,
   private val eeProperties: EeProperties,
   private val reportingService: UsageReportingService,
+  private val schedulingManager: SchedulingManager
 ) : Logging {
-  companion object {
-    private val scheduledTasks = ConcurrentHashMap<String, ScheduledFuture<*>>()
-
-    fun cancelAll() {
-      // Create a thread-safe copy of the keys
-      val idsToCancel = ArrayList(scheduledTasks.keys)
-      logger().debug("Cancelling all ${idsToCancel.size} scheduled tasks")
-      idsToCancel.forEach { cancelTask(it) }
-    }
-
-    fun cancelTask(id: String) {
-      scheduledTasks[id]?.cancel(true)
-      scheduledTasks.remove(id)
-    }
-  }
-
-  // Store the ID of the task scheduled at startup for potential cancellation
-  private var startupTaskId: String? = null
-
   @EventListener(ApplicationReadyEvent::class)
   fun scheduleReporting() {
     if (!eeProperties.scheduledReportingEnabled) {
       logger.info("Scheduled reporting is disabled, skipping scheduling")
       return
     }
-    startupTaskId = scheduleTask()
-    logger.debug("Scheduled reporting task with ID: $startupTaskId")
+    scheduleTask()
+    logger.debug("Scheduled reporting task")
   }
 
-  fun scheduleTask(): String {
+  fun scheduleTask() {
     val period = Duration.ofMillis(eeProperties.reportUsageFixedDelayInMs)
     val runnable = {
       logger.debug("Reporting usage periodically")
       reportingService.reportIfNeeded()
     }
-    val future = taskScheduler.scheduleAtFixedRate(runnable, period)
-    val id = UUID.randomUUID().toString()
-    scheduledTasks[id] = future
-    return id
-  }
 
-  @PreDestroy
-  fun cancelAll() {
-    Companion.cancelAll()
+    schedulingManager.scheduleWithFixedDelay(runnable, period)
   }
 }
